rename: @/components/common/nav-bar/mobile 디렉토리 생성
